final-pjt-back/requirements.txt
asgiref==3.7.2
asttokens==2.4.0
attrs==23.1.0
axios==0.4.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
branca==0.7.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.4
comm==0.1.3
ConfigArgParse==1.7
contourpy==1.1.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.6.7
decorator==5.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.4
django-allauth==0.58.2
django-bootstrap-v5==1.0.2
django-cors-headers==4.2.0
django-environ==0.11.2
django-extensions==3.1.3
django-seed==0.2.2
djangorestframework==3.14.0
environ==1.0
executing==1.2.0
Faker==8.1.1
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
folium==0.15.0
fonttools==4.43.1
gevent==23.9.1
geventhttpclient==2.0.11
greenlet==3.0.1
h11==0.14.0
idna==3.4
importlib-metadata==2.1.1
ipykernel==6.25.0
ipython==7.22.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.5
locust==2.18.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.3
mdurl==0.1.2
msgpack==1.0.7
nest-asyncio==1.5.6
numpy==1.26.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.2
parso==0.8.2
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.9.1
prompt-toolkit==3.0.18
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydot==1.4.2
Pygments==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
roundrobin==0.0.4
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.14.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
threadpoolctl==3.2.0
tornado==6.3.2
traitlets==5.0.5
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.5
Werkzeug==3.0.1
wsproto==1.2.0
zipp==3.4.1
zope.event==5.0
zope.interface==6.1
